projects/HashScope/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.20
alembic==1.12.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
web3==6.9.0
eth-account==0.9.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pytest==7.4.2
httpx==0.24.1
